EDGECLOUD-591: Follow-up changes (#1036)
* EDGECLOUD-591: Follow-up changes

* Add vault ssh cookbook to download CA from respective setup vault

* upgrade fixes

+++ vmlayer/ssh.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/mobiledgex/edge-cloud-infra/infracommon"
 	"github.com/mobiledgex/edge-cloud/cloud-resource-manager/platform"
+	"github.com/mobiledgex/edge-cloud/cloud-resource-manager/platform/pc"
 	"github.com/mobiledgex/edge-cloud/cloudcommon"
 	"github.com/mobiledgex/edge-cloud/edgeproto"
 	"github.com/mobiledgex/edge-cloud/log"
@@ -138,8 +139,9 @@ func (vp *VMProperties) GetSSHClientFromIPAddr(ctx context.Context, ipaddr strin
 	vp.sshKey.Mux.Unlock()
 
 	if vp.sshKey.UseMEXPrivateKey {
-		auth = ssh.Auth{RawKeys: [][]byte{}}
-		auth.RawKeys = append(auth.RawKeys, []byte(vp.sshKey.MEXPrivateKey))
+		auth = ssh.Auth{RawKeys: [][]byte{
+			[]byte(vp.sshKey.MEXPrivateKey),
+		}}
 	}
 
 	gwhost, gwport := vp.GetCloudletCRMGatewayIPAndPort()
@@ -239,7 +241,7 @@ func (v *VMPlatform) GetAllCloudletVMs(ctx context.Context, caches *platform.Cac
 	pfName := v.GetPlatformVMName(v.VMProperties.CommonPf.PlatformConfig.CloudletKey)
 	client, err := v.GetSSHClientForServer(ctx, pfName, v.VMProperties.GetCloudletExternalNetwork())
 	if err != nil {
-		return nil, err
+		log.SpanLog(ctx, log.DebugLevelInfra, "error getting ssh client", "vm", pfName, "err", err)
 	}
 	cloudletVMs = append(cloudletVMs, VMAccess{
 		Name:   pfName,
@@ -251,7 +253,7 @@ func (v *VMPlatform) GetAllCloudletVMs(ctx context.Context, caches *platform.Cac
 	sharedRootLBName := v.VMProperties.SharedRootLBName
 	sharedlbclient, err := v.GetSSHClientForServer(ctx, sharedRootLBName, v.VMProperties.GetCloudletExternalNetwork())
 	if err != nil {
-		return nil, err
+		log.SpanLog(ctx, log.DebugLevelInfra, "error getting ssh client", "vm", sharedRootLBName, "err", err)
 	}
 	cloudletVMs = append(cloudletVMs, VMAccess{
 		Name:   sharedRootLBName,
@@ -272,7 +274,7 @@ func (v *VMPlatform) GetAllCloudletVMs(ctx context.Context, caches *platform.Cac
 				rootLBName := v.VMProperties.GetRootLBNameForCluster(ctx, clusterInst)
 				dedicatedlbclient, err = v.GetSSHClientForServer(ctx, rootLBName, v.VMProperties.GetCloudletExternalNetwork())
 				if err != nil {
-					return nil, err
+					log.SpanLog(ctx, log.DebugLevelInfra, "error getting ssh client", "vm", rootLBName, "err", err)
 				}
 				cloudletVMs = append(cloudletVMs, VMAccess{
 					Name:   rootLBName,
@@ -290,7 +292,7 @@ func (v *VMPlatform) GetAllCloudletVMs(ctx context.Context, caches *platform.Cac
 			masterNode := GetClusterMasterName(ctx, clusterInst)
 			masterIP, err := v.GetIPFromServerName(ctx, v.VMProperties.GetCloudletMexNetwork(), GetClusterSubnetName(ctx, clusterInst), masterNode)
 			if err != nil {
-				return nil, err
+				log.SpanLog(ctx, log.DebugLevelInfra, "error getting ssh client", "vm", masterNode, "err", err)
 			}
 			masterClient, err := lbClient.AddHop(masterIP.ExternalAddr, 22)
 			cloudletVMs = append(cloudletVMs, VMAccess{
@@ -302,7 +304,7 @@ func (v *VMPlatform) GetAllCloudletVMs(ctx context.Context, caches *platform.Cac
 				clusterNode := GetClusterNodeName(ctx, clusterInst, nn)
 				nodeIP, err := v.GetIPFromServerName(ctx, v.VMProperties.GetCloudletMexNetwork(), GetClusterSubnetName(ctx, clusterInst), clusterNode)
 				if err != nil {
-					return nil, err
+					log.SpanLog(ctx, log.DebugLevelInfra, "error getting ssh client", "vm", clusterNode, "err", err)
 				}
 				nodeClient, err := lbClient.AddHop(nodeIP.ExternalAddr, 22)
 				cloudletVMs = append(cloudletVMs, VMAccess{
@@ -316,33 +318,36 @@ func (v *VMPlatform) GetAllCloudletVMs(ctx context.Context, caches *platform.Cac
 	return cloudletVMs, nil
 }
 
-func (v *VMPlatform) UpgradeFuncHandleSSHKeys(ctx context.Context, vaultConfig *vault.Config, caches *platform.Caches) (map[string]string, error) {
-	log.SpanLog(ctx, log.DebugLevelInfra, "Upgrade Vms to use Vault SSH signed keys")
-	// Pull private key from Vault
-	mexKey, err := infracommon.GetMEXKeyFromVault(vaultConfig)
-	if err != nil {
-		return nil, err
-	}
-	v.VMProperties.sshKey.MEXPrivateKey = mexKey.PrivateKey
+func GetVaultCAScript(vaultConfig *vault.Config) string {
+	return fmt.Sprintf(`
+#!/bin/bash
 
-	// Set SSH client to use mex private key
-	v.VMProperties.sshKey.UseMEXPrivateKey = true
-	fixVMs, err := v.GetAllCloudletVMs(ctx, caches)
-	if err != nil {
-		return nil, err
-	}
+die() {
+        echo "ERROR: $*" >&2
+        exit 2
+}
 
-	setupVaultCAScript := fmt.Sprintf(`
 curl %s/v1/ssh/public_key | sudo tee /etc/ssh/trusted-user-ca-keys.pem
+[[ $? -ne 0 ]] && die "failed to get CA cert from vault"
+sudo grep "ssh-rsa" /etc/ssh/trusted-user-ca-keys.pem
+[[ $? -ne 0 ]] && die "invalid CA cert from vault"
+
 echo 'TrustedUserCAKeys /etc/ssh/trusted-user-ca-keys.pem' | sudo tee -a /etc/ssh/sshd_config
+sudo systemctl reload ssh
+[[ $? -ne 0 ]] && die "failed to reload ssh"
+
 rm -f id_rsa_mex
 echo "" > .ssh/authorized_keys
-sudo systemctl reload ssh
-`, vaultConfig.Addr)
 
-	setupK8sJoinSvcScript := `
+echo "Done setting up vault ssh"
+`, vaultConfig.Addr)
+}
+
+func GetVaultCAScriptForMasterNode(vaultConfig *vault.Config) string {
+	k8sJoinSvcScript := `
 mkdir -p /var/tmp/k8s-join
 cp /tmp/k8s-join-cmd.tmp /var/tmp/k8s-join/k8s-join-cmd
+[[ $? -ne 0 ]] && die "failed to copy k8s-join-cmd.tmp file"
 chown ubuntu:ubuntu /var/tmp/k8s-join/k8s-join-cmd
 
 sudo tee /etc/systemd/system/k8s-join.service <<'EOT'
@@ -358,19 +363,56 @@ WantedBy=multi-user.target
 EOT
 
 sudo systemctl enable k8s-join
-sudo systemctl start k8s-join`
+[[ $? -ne 0 ]] && die "failed to enable k8s-join service"
+sudo systemctl start k8s-join
+[[ $? -ne 0 ]] && die "failed to start k8s-join service"
+
+echo "Done setting k8s-join service"
+`
+	vaultCAScript := GetVaultCAScript(vaultConfig)
+	return vaultCAScript + k8sJoinSvcScript
+}
+
+func ExecuteUpgradeScript(ctx context.Context, vmName string, client ssh.Client, script string) error {
+	log.SpanLog(ctx, log.DebugLevelInfra, "execute upgrade script", "vmName", vmName)
+	err := pc.WriteFile(client, "upgradeCRMVault.sh", script, "upgrade script", pc.NoSudo)
+	if err != nil {
+		log.SpanLog(ctx, log.DebugLevelInfra, "failed to copy script", "err", err)
+		return err
+	}
+	// Execute script
+	out, err := client.Output("bash upgradeCRMVault.sh")
+	if err != nil {
+		log.SpanLog(ctx, log.DebugLevelInfra, "failed to fix vm", "out", out, "err", err)
+		return err
+	}
+	return nil
+}
+
+func (v *VMPlatform) UpgradeFuncHandleSSHKeys(ctx context.Context, vaultConfig *vault.Config, caches *platform.Caches) (map[string]string, error) {
+	log.SpanLog(ctx, log.DebugLevelInfra, "Upgrade Vms to use Vault SSH signed keys")
+	// Set SSH client to use mex private key
+	v.VMProperties.sshKey.UseMEXPrivateKey = true
+	fixVMs, err := v.GetAllCloudletVMs(ctx, caches)
+	if err != nil {
+		return nil, err
+	}
 
 	for _, vm := range fixVMs {
-		script := setupVaultCAScript
+		if vm.Client == nil {
+			log.SpanLog(ctx, log.DebugLevelInfra, "missing ssh client", "vm", vm.Name)
+			continue
+		}
+		script := ""
 		if vm.Role == RoleMaster {
 			// Start k8s-join webserver
-			script += setupK8sJoinSvcScript
+			script = GetVaultCAScriptForMasterNode(vaultConfig)
+		} else {
+			script = GetVaultCAScript(vaultConfig)
 		}
-		// Execute script
-		out, err := vm.Client.Output(script)
+		err = ExecuteUpgradeScript(ctx, vm.Name, vm.Client, script)
 		if err != nil {
-			// log the error and continue fixing other VMs
-			log.SpanLog(ctx, log.DebugLevelInfra, "failed to fix vm", "out", out, "err", err)
+			// continue fixing other VMs
 			continue
 		}
 	}
@@ -384,9 +426,13 @@ sudo systemctl start k8s-join`
 	}
 	results := make(map[string]string)
 	for _, vm := range fixVMs {
+		if vm.Client == nil {
+			results[vm.Name] = "failed to get ssh client"
+			continue
+		}
 		_, err = vm.Client.Output("hostname")
 		if err != nil {
-			results[vm.Name] = fmt.Sprintf("failed with err %v", err)
+			results[vm.Name] = fmt.Sprintf("failed with error: %v", err)
 			continue
 		}
 		results[vm.Name] = "fixed"
